Fix error while recording with UMI Max
The code has a small refactor, and add's a fallback recording option
without the resolution argument if the one with "--size w x h" fails.

UMI Max with Android 7, has a 1920x1080 screen, but trying to
record with this resolution set on "adb" "--size" argument, results in
error, saying that only 1920x1088 is supported.
(note the "1088" versus "1080")